ai-news-aggregator/requirements.txt
mysql-connector-python==8.0.33
praw==7.7.1
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
